feature: look for session_ttl in okta section of AWS config
If not provided by a more immediate means (i.e. the `-session-ttl` flag or `AWS_SESSION_TTL` environment variable), look for it in the `[okta]` section of the AWS config under `session_ttl`.

## Why not allow it to be set per-profile?

Due to the way the Okta session is cached, per-profile Okta session TTLs don't really make sense.

The issue is: If there's an unexpired session it is used regardless of TTL. This leads to potentially unexpected behavior when switching between profiles with different TTLs.

For example: You call `aws-okta` on a profile with a short TTL. Then, you call `aws-okta` on a profile with a long TTL. You are then surprised to find you need to reauth sooner than expected because the short TTL was still in effect.

To avoid this confusion, we set the Okta session TTL in the `[okta]` section of the AWS config. So it's clearer that this setting is not going to readily change between profiles.

The `AWS_SESSION_TTL` environment variable and `-session-ttl` flag both will override the AWS config value.

Note: The behavior described isn't specific to this change. The same thing happens when changing the value with `AWS_SESSION_TTL` or `-session-ttl`.